+++ custom_components/edgeos/component/models/edge_os_device_data.py
@@ -40,7 +40,7 @@ class EdgeOSDeviceData:
     def last_activity_in_seconds(self) -> float:
         now = datetime.now().timestamp()
 
-        diff = now - self.last_activity
+        diff = int(now) - self.last_activity
         last_activity_in_seconds = timedelta(seconds=diff).total_seconds()
 
         return last_activity_in_seconds