+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Add IP address to status switch of its interface [#71](https://github.com/elad-bar/ha-edgeos/issues/71)
 - Constants clean up
 - Add ability to set the interval to update data and entities separately, to update use the `edgeos.update_configuration` service and set the number of seconds per `update_api_interval` and/or `update_entities_interval` parameters, defaults are API: 30 seconds, Entities: 1 second
+- Improved logic of service `edgeos.update_configuration`
+- Add to store debug data HA data that is being used to generate HA components
+- Add openvpn as supported interface type
 
 ## 2.0.3